requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-telegram-bot==21.6
httpx==0.27.2
python-dotenv==1.0.1
websockets==13.1
psycopg2-binary==2.9.9
